notion2docx/scripts/requirements.txt
requests>=2.0.0
pywin32>=306
latex2mathml >=3.18.1
pypandoc>=1.13
python-docx>=1.2.0
